test(studio-api): split server route regression tests
