+++ packages/studio/src/api/server.test.ts
@@ -1,449 +1,55 @@
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { access, mkdir, mkdtemp, readFile, rm, writeFile } from "node:fs/promises";
+import { describe, expect, it, vi } from "vitest";
+import { access, mkdir, readFile, writeFile } from "node:fs/promises";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
-
-const schedulerStartMock = vi.fn<() => Promise<void>>();
-const initBookMock = vi.fn();
-const runRadarMock = vi.fn();
-const reviseDraftMock = vi.fn();
-const resyncChapterArtifactsMock = vi.fn();
-const writeNextChapterMock = vi.fn();
-const rollbackToChapterMock = vi.fn();
-const saveChapterIndexMock = vi.fn();
-const loadChapterIndexMock = vi.fn();
-const loadBookConfigMock = vi.fn();
-const createLLMClientMock = vi.fn(() => ({}));
-const chatCompletionMock = vi.fn();
-const loadProjectConfigMock = vi.fn();
-const pipelineConfigs: unknown[] = [];
-const processProjectInteractionInputMock = vi.fn();
-const processProjectInteractionRequestMock = vi.fn();
-const createInteractionToolsFromDepsMock = vi.fn(() => ({}));
-const loadProjectSessionMock = vi.fn();
-const resolveSessionActiveBookMock = vi.fn();
-const runAgentSessionMock = vi.fn();
-const createAndPersistBookSessionMock = vi.fn();
-const loadBookSessionMock = vi.fn();
-const persistBookSessionMock = vi.fn();
-const appendBookSessionMessageMock = vi.fn();
-const appendManualSessionMessagesMock = vi.fn();
-const renameBookSessionMock = vi.fn();
-const deleteBookSessionMock = vi.fn();
-const migrateBookSessionMock = vi.fn();
-const resolveServiceModelMock = vi.fn();
-const loadSecretsMock = vi.fn();
-const saveSecretsMock = vi.fn();
-const getServiceApiKeyMock = vi.fn();
-type ServicePresetMock = {
-  providerFamily: "openai" | "anthropic";
-  baseUrl: string;
-  modelsBaseUrl?: string;
-  knownModels: string[];
-};
-const SERVICE_PRESETS_MOCK: Record<string, ServicePresetMock> = {
-  openai: { providerFamily: "openai", baseUrl: "https://api.openai.com/v1", modelsBaseUrl: "https://api.openai.com/v1", knownModels: [] as string[] },
-  anthropic: { providerFamily: "anthropic", baseUrl: "https://api.anthropic.com", modelsBaseUrl: "https://api.anthropic.com", knownModels: [] as string[] },
-  minimax: { providerFamily: "anthropic", baseUrl: "https://api.minimaxi.com/anthropic", modelsBaseUrl: "https://api.minimaxi.com/anthropic", knownModels: [] as string[] },
-  bailian: { providerFamily: "anthropic", baseUrl: "https://dashscope.aliyuncs.com/apps/anthropic", modelsBaseUrl: "https://dashscope.aliyuncs.com/compatible-mode/v1", knownModels: [] as string[] },
-  google: { providerFamily: "openai", baseUrl: "https://generativelanguage.googleapis.com/v1beta/openai", modelsBaseUrl: "https://generativelanguage.googleapis.com/v1beta/openai", knownModels: [] as string[] },
-  ollama: { providerFamily: "openai", baseUrl: "http://localhost:11434/v1", modelsBaseUrl: "http://localhost:11434/v1", knownModels: [] as string[] },
-  custom: { providerFamily: "openai", baseUrl: "", knownModels: [] as string[] },
-};
-const resolveServicePresetMock = vi.fn((service: string) => SERVICE_PRESETS_MOCK[service]);
-const resolveServiceProviderFamilyMock = vi.fn((service: string) => resolveServicePresetMock(service)?.providerFamily);
-const resolveServiceModelsBaseUrlMock = vi.fn((service: string) => {
-  const preset = SERVICE_PRESETS_MOCK[service];
-  return preset?.modelsBaseUrl ?? preset?.baseUrl;
-});
-const listModelsForServiceMock = vi.fn(async (service: string, apiKey?: string, liveBaseUrl?: string) => {
-  const preset = resolveServicePresetMock(service);
-  if (!preset) return [];
-  if (preset.knownModels.length > 0) {
-    return preset.knownModels.map((id) => ({ id, name: id, reasoning: false, contextWindow: 0 }));
-  }
-  const modelsBaseUrl = liveBaseUrl ?? resolveServiceModelsBaseUrlMock(service);
-  const allowsNoKey = Boolean(modelsBaseUrl?.startsWith("http://localhost") || modelsBaseUrl?.startsWith("http://127.0.0.1"));
-  if ((!apiKey && !allowsNoKey) || !modelsBaseUrl) return [];
-  const res = await fetch(`${modelsBaseUrl.replace(/\/$/, "")}/models`, {
-    headers: apiKey ? { Authorization: `Bearer ${apiKey}` } : {},
-    signal: AbortSignal.timeout(10_000),
-  });
-  if (!res.ok) return [];
-  const json = await res.json() as { data?: Array<{ id: string }> };
-  return (json.data ?? []).map((model) => ({
-    id: model.id,
-    name: model.id,
-    reasoning: false,
-    contextWindow: 0,
-  }));
-});
-const endpointIdsByGroup = {
-  overseas: ["anthropic", "google", "mistral", "openai", "xai"],
-  china: [
-    "ai360", "baichuan", "bailian", "deepseek", "hunyuan", "internlm", "longcat",
-    "minimax", "moonshot", "sensenova", "spark", "stepfun", "tencentcloud",
-    "volcengine", "wenxin", "xiaomimimo", "zeroone", "zhipu",
-  ],
-  aggregator: ["giteeai", "infiniai", "modelscope", "newapi", "openrouter", "ppio", "qiniu", "siliconcloud"],
-  local: ["githubCopilot", "ollama"],
-  codingPlan: [
-    "astronCodingPlan", "bailianCodingPlan", "glmCodingPlan", "kimiCodingPlan", "kimicode",
-    "minimaxCodingPlan", "opencodeCodingPlan", "volcengineCodingPlan",
-  ],
-} as const;
-const endpointMocks = [
-  ...Object.entries(endpointIdsByGroup).flatMap(([group, ids]) => ids.map((id) => ({
-    id,
-    label: id,
-    group,
-    ...(id === "google" ? { checkModel: "gemini-2.5-flash" } : {}),
-    ...(id === "minimax" ? { checkModel: "MiniMax-M2.7" } : {}),
-    ...(id === "ollama" ? { checkModel: "llama3.2:3b" } : {}),
-    ...(id === "volcengine" ? { checkModel: "doubao-lite-32k" } : {}),
-    models: [
-      { id: `${id}-model`, maxOutput: 4096, contextWindowTokens: 32768, enabled: true },
-      { id: `${id}-disabled`, maxOutput: 4096, contextWindowTokens: 32768, enabled: false },
-    ],
-  }))),
-  { id: "custom", label: "自定义端点", models: [] },
-];
-const getAllEndpointsMock = vi.fn(() => endpointMocks);
-const probeModelsFromUpstreamMock = vi.fn(async () => [
-  { id: "custom-model", name: "custom-model", contextWindow: 0 },
-]);
-
-const logger = {
-  child: () => logger,
-  info: vi.fn(),
-  warn: vi.fn(),
-  error: vi.fn(),
-};
-
-vi.mock("@actalk/inkos-core", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("@actalk/inkos-core")>();
-
-  class MockSessionAlreadyMigratedError extends Error {
-    constructor(message = "Session already migrated") {
-      super(message);
-      this.name = "SessionAlreadyMigratedError";
-    }
-  }
-
-  class MockStateManager {
-    constructor(private readonly root: string) {}
-
-    async listBooks(): Promise<string[]> {
-      return [];
-    }
-
-    async loadBookConfig(bookId?: string): Promise<never> {
-      return await loadBookConfigMock(bookId) as never;
-    }
-
-    async loadChapterIndex(bookId: string): Promise<[]> {
-      return (await loadChapterIndexMock(bookId)) as [];
-    }
-
-    async saveChapterIndex(bookId: string, index: unknown): Promise<void> {
-      await saveChapterIndexMock(bookId, index);
-    }
-
-    async rollbackToChapter(bookId: string, chapterNumber: number): Promise<number[]> {
-      return (await rollbackToChapterMock(bookId, chapterNumber)) as number[];
-    }
-
-    async getNextChapterNumber(_bookId?: string): Promise<number> {
-      return 1;
-    }
-
-    async ensureControlDocuments(): Promise<void> {
-      // no-op in tests
-    }
-
-    bookDir(id: string): string {
-      return join(this.root, "books", id);
-    }
-  }
-
-  class MockPipelineRunner {
-    constructor(config: unknown) {
-      pipelineConfigs.push(config);
-    }
-
-    initBook = initBookMock;
-    runRadar = runRadarMock;
-    reviseDraft = reviseDraftMock;
-    resyncChapterArtifacts = resyncChapterArtifactsMock;
-    writeNextChapter = writeNextChapterMock;
-  }
-
-  class MockScheduler {
-    private running = false;
-
-    constructor(_config: unknown) {}
-
-    async start(): Promise<void> {
-      this.running = true;
-      await schedulerStartMock();
-    }
-
-    stop(): void {
-      this.running = false;
-    }
-
-    get isRunning(): boolean {
-      return this.running;
-    }
-  }
-
-  return {
-    StateManager: MockStateManager,
-    PipelineRunner: MockPipelineRunner,
-    Scheduler: MockScheduler,
-    createLLMClient: createLLMClientMock,
-    createLogger: vi.fn(() => logger),
-    computeAnalytics: vi.fn(() => ({})),
-    isSafeBookId: actual.isSafeBookId,
-    normalizePlatformOrOther: actual.normalizePlatformOrOther,
-    chatCompletion: chatCompletionMock,
-    loadProjectConfig: loadProjectConfigMock,
-    processProjectInteractionInput: processProjectInteractionInputMock,
-    processProjectInteractionRequest: processProjectInteractionRequestMock,
-    createInteractionToolsFromDeps: createInteractionToolsFromDepsMock,
-    loadProjectSession: loadProjectSessionMock,
-    resolveSessionActiveBook: resolveSessionActiveBookMock,
-    runAgentSession: runAgentSessionMock,
-    buildAgentSystemPrompt: vi.fn(() => "You are helpful."),
-    createAndPersistBookSession: createAndPersistBookSessionMock,
-    loadBookSession: loadBookSessionMock,
-    persistBookSession: persistBookSessionMock,
-    appendBookSessionMessage: appendBookSessionMessageMock,
-    appendManualSessionMessages: appendManualSessionMessagesMock,
-    isNewLayoutBook: vi.fn(async () => false),
-    renameBookSession: renameBookSessionMock,
-    deleteBookSession: deleteBookSessionMock,
-    migrateBookSession: migrateBookSessionMock,
-    SessionAlreadyMigratedError: MockSessionAlreadyMigratedError,
-    resolveServicePreset: resolveServicePresetMock,
-    resolveServiceProviderFamily: resolveServiceProviderFamilyMock,
-    resolveServiceModelsBaseUrl: resolveServiceModelsBaseUrlMock,
-    resolveServiceModel: resolveServiceModelMock,
-    isApiKeyOptionalForEndpoint: actual.isApiKeyOptionalForEndpoint,
-    loadSecrets: loadSecretsMock,
-    saveSecrets: saveSecretsMock,
-    getServiceApiKey: getServiceApiKeyMock,
-    listModelsForService: listModelsForServiceMock,
-    getAllEndpoints: getAllEndpointsMock,
-    probeModelsFromUpstream: probeModelsFromUpstreamMock,
-    fetchWithProxy: vi.fn((input: Parameters<typeof fetch>[0], init?: RequestInit) => fetch(input, init)),
-    GLOBAL_ENV_PATH: join(tmpdir(), "inkos-global.env"),
-  };
-});
-
-const projectConfig = {
-  name: "studio-test",
-  version: "0.1.0",
-  language: "zh",
-  llm: {
-    provider: "openai",
-    baseUrl: "https://api.example.com/v1",
-    apiKey: "sk-test",
-    model: "gpt-5.4",
-    temperature: 0.7,
-    maxTokens: 4096,
-    stream: false,
-  },
-  daemon: {
-    schedule: {
-      radarCron: "0 */6 * * *",
-      writeCron: "*/15 * * * *",
-    },
-    maxConcurrentBooks: 1,
-    chaptersPerCycle: 1,
-    retryDelayMs: 30000,
-    cooldownAfterChapterMs: 0,
-    maxChaptersPerDay: 50,
-  },
-  modelOverrides: {},
-  notify: [],
-} as const;
-
-function cloneProjectConfig() {
-  return structuredClone(projectConfig);
-}
+import {
+  root,
+  schedulerStartMock,
+  initBookMock,
+  runRadarMock,
+  reviseDraftMock,
+  resyncChapterArtifactsMock,
+  writeNextChapterMock,
+  rollbackToChapterMock,
+  saveChapterIndexMock,
+  loadChapterIndexMock,
+  loadBookConfigMock,
+  createLLMClientMock,
+  chatCompletionMock,
+  loadProjectConfigMock,
+  pipelineConfigs,
+  processProjectInteractionInputMock,
+  processProjectInteractionRequestMock,
+  createInteractionToolsFromDepsMock,
+  loadProjectSessionMock,
+  resolveSessionActiveBookMock,
+  runAgentSessionMock,
+  createAndPersistBookSessionMock,
+  loadBookSessionMock,
+  persistBookSessionMock,
+  appendBookSessionMessageMock,
+  appendManualSessionMessagesMock,
+  renameBookSessionMock,
+  deleteBookSessionMock,
+  migrateBookSessionMock,
+  resolveServiceModelMock,
+  loadSecretsMock,
+  saveSecretsMock,
+  getServiceApiKeyMock,
+  SERVICE_PRESETS_MOCK,
+  resolveServicePresetMock,
+  resolveServiceProviderFamilyMock,
+  resolveServiceModelsBaseUrlMock,
+  listModelsForServiceMock,
+  endpointIdsByGroup,
+  endpointMocks,
+  getAllEndpointsMock,
+  probeModelsFromUpstreamMock,
+  projectConfig,
+  cloneProjectConfig
+} from "./__tests__/server-test-utils";
 
 describe("createStudioServer daemon lifecycle", () => {
-  let root: string;
-
-  beforeEach(async () => {
-    root = await mkdtemp(join(tmpdir(), "inkos-studio-server-"));
-    await writeFile(join(root, "inkos.json"), JSON.stringify(projectConfig, null, 2), "utf-8");
-    schedulerStartMock.mockReset();
-    initBookMock.mockReset();
-    runRadarMock.mockReset();
-    reviseDraftMock.mockReset();
-    resyncChapterArtifactsMock.mockReset();
-    writeNextChapterMock.mockReset();
-    rollbackToChapterMock.mockReset();
-    saveChapterIndexMock.mockReset();
-    loadChapterIndexMock.mockReset();
-    loadBookConfigMock.mockReset();
-    await mkdir(join(root, "books", "demo-book", "chapters"), { recursive: true });
-    await writeFile(join(root, "books", "demo-book", "chapters", "0003_Demo.md"), "# Demo\n\nBody", "utf-8");
-    runRadarMock.mockResolvedValue({
-      marketSummary: "Fresh market summary",
-      recommendations: [],
-    });
-    reviseDraftMock.mockResolvedValue({
-      chapterNumber: 3,
-      wordCount: 1800,
-      fixedIssues: ["focus restored"],
-      applied: true,
-      status: "ready-for-review",
-    });
-    resyncChapterArtifactsMock.mockResolvedValue({
-      chapterNumber: 3,
-      title: "Synced Chapter",
-      wordCount: 1800,
-      revised: false,
-      status: "ready-for-review",
-      auditResult: { passed: true, issues: [], summary: "synced" },
-    });
-    writeNextChapterMock.mockResolvedValue({
-      chapterNumber: 3,
-      title: "Rewritten Chapter",
-      wordCount: 1800,
-      revised: false,
-      status: "ready-for-review",
-      auditResult: { passed: true, issues: [], summary: "rewritten" },
-    });
-    createLLMClientMock.mockReset();
-    createLLMClientMock.mockReturnValue({});
-    chatCompletionMock.mockReset();
-    chatCompletionMock.mockResolvedValue({
-      content: "pong",
-      usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-    });
-    loadProjectConfigMock.mockReset();
-    processProjectInteractionInputMock.mockReset();
-    processProjectInteractionRequestMock.mockReset();
-    createInteractionToolsFromDepsMock.mockReset();
-    loadProjectSessionMock.mockReset();
-    resolveSessionActiveBookMock.mockReset();
-    createInteractionToolsFromDepsMock.mockReturnValue({});
-    processProjectInteractionRequestMock.mockResolvedValue({
-      request: { intent: "create_book" },
-      session: {
-        sessionId: "session-structured",
-        projectRoot: root,
-        activeBookId: "new-book",
-        automationMode: "semi",
-        messages: [],
-        events: [],
-      },
-      details: {
-        bookId: "new-book",
-        outputPath: join(root, "books", "demo-book", "demo-book.txt"),
-        chaptersExported: 2,
-      },
-    });
-    loadProjectSessionMock.mockResolvedValue({
-      sessionId: "session-1",
-      projectRoot: root,
-      automationMode: "semi",
-      messages: [],
-    });
-    resolveSessionActiveBookMock.mockResolvedValue(undefined);
-    loadProjectConfigMock.mockImplementation(async () => {
-      const raw = JSON.parse(await readFile(join(root, "inkos.json"), "utf-8")) as Record<string, unknown>;
-      return {
-        ...cloneProjectConfig(),
-        ...raw,
-        llm: {
-          ...cloneProjectConfig().llm,
-          ...((raw.llm ?? {}) as Record<string, unknown>),
-        },
-        daemon: {
-          ...cloneProjectConfig().daemon,
-          ...((raw.daemon ?? {}) as Record<string, unknown>),
-        },
-        modelOverrides: (raw.modelOverrides ?? {}) as Record<string, unknown>,
-        notify: (raw.notify ?? []) as unknown[],
-      };
-    });
-    loadChapterIndexMock.mockResolvedValue([]);
-    loadBookConfigMock.mockResolvedValue({
-      id: "demo-book",
-      title: "Demo Book",
-      platform: "qidian",
-      genre: "xuanhuan",
-      status: "active",
-      targetChapters: 100,
-      chapterWordCount: 3000,
-      createdAt: "2026-04-12T00:00:00.000Z",
-      updatedAt: "2026-04-12T00:00:00.000Z",
-    });
-    saveChapterIndexMock.mockResolvedValue(undefined);
-    rollbackToChapterMock.mockResolvedValue([]);
-    pipelineConfigs.length = 0;
-    runAgentSessionMock.mockReset();
-    createAndPersistBookSessionMock.mockReset();
-    loadBookSessionMock.mockReset();
-    persistBookSessionMock.mockReset();
-    appendBookSessionMessageMock.mockReset();
-    appendManualSessionMessagesMock.mockReset();
-    renameBookSessionMock.mockReset();
-    deleteBookSessionMock.mockReset();
-    migrateBookSessionMock.mockReset();
-    resolveServiceModelMock.mockReset();
-    loadSecretsMock.mockReset();
-    saveSecretsMock.mockReset();
-    getServiceApiKeyMock.mockReset();
-    resolveServicePresetMock.mockClear();
-    resolveServiceProviderFamilyMock.mockClear();
-    resolveServiceModelsBaseUrlMock.mockClear();
-    listModelsForServiceMock.mockClear();
-    getAllEndpointsMock.mockClear();
-    probeModelsFromUpstreamMock.mockClear();
-    // Default BookSession for agent tests
-    const defaultBookSession = {
-      sessionId: "agent-session-1",
-      bookId: "demo-book",
-      title: null,
-      messages: [],
-      events: [],
-      draftRounds: [],
-      createdAt: 1,
-      updatedAt: 1,
-    };
-    createAndPersistBookSessionMock.mockResolvedValue(defaultBookSession);
-    loadBookSessionMock.mockResolvedValue(defaultBookSession);
-    persistBookSessionMock.mockResolvedValue(undefined);
-    appendBookSessionMessageMock.mockImplementation(
-      (session: unknown, _msg: unknown) => session,
-    );
-    appendManualSessionMessagesMock.mockResolvedValue(undefined);
-    renameBookSessionMock.mockResolvedValue(null);
-    deleteBookSessionMock.mockResolvedValue(undefined);
-    migrateBookSessionMock.mockImplementation(async (_root: string, _sessionId: string, bookId: string) => ({
-      ...defaultBookSession,
-      bookId,
-    }));
-    runAgentSessionMock.mockResolvedValue({
-      responseText: "Agent response.",
-      messages: [],
-    });
-    loadSecretsMock.mockResolvedValue({ services: {} });
-    saveSecretsMock.mockResolvedValue(undefined);
-    getServiceApiKeyMock.mockResolvedValue(undefined);
-  });
-
-  afterEach(async () => {
-    await rm(root, { recursive: true, force: true });
-    await rm(join(tmpdir(), "inkos-global.env"), { force: true });
-  });
-
   it("uses the real core bookId validator in the Studio safety mock", async () => {
     const { isSafeBookId } = await import("@actalk/inkos-core");
